.github/workflows/A380_AI_Crew_GateToGate/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "a380-ai-crew"
version = "0.2.0"
description = "A380 AI Crew Commander (Gate-to-Gate) for MSFS2024 + FBW A380X"
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
  "PyYAML>=6.0.1",
  "SimConnect==0.4.26",
  "pythonnet>=3.0.5",
  "clr_loader>=0.2.10",
  "cffi>=2.0.0",
  "pycparser>=2.22",
  "requests>=2.31.0",
]

[project.scripts]
a380-ai-crew = "a380_ai_crew.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
